+++ test_e2e.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+set -eux
+
+go test ./tests_e2e/...
